+++ lib/background_patterns.dart
@@ -1,3 +1,4 @@
 library background_patterns;
 
 export 'src/pattern_container.dart';
+export 'src/shapes/shape_configs.dart';