+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Initial setup of a background container
+
+## 0.0.3
+
+* Shapes in pattern made configurable